+++ CHANGELOG.md
@@ -65,7 +65,8 @@
 * Add Hub search behavior for LoC source; fix search results table. Refs [UILD-688].
 * Add Hubs actions menu. Refs [UILD-693].
 * Add initial workspace profile settings editing. Refs [UILD-685].
-* Refactor Edit page to support multiple resource types. Add Hub create page. Refs [UILD-694]. 
+* Refactor Edit page to support multiple resource types. Add Hub create page. Refs [UILD-694].
+* Add workspace profile settings components. Refs [UILD-696].
 
 [UILD-552]:https://folio-org.atlassian.net/browse/UILD-552
 [UILD-544]:https://folio-org.atlassian.net/browse/UILD-544
@@ -132,6 +133,7 @@
 [UILD-693]:https://folio-org.atlassian.net/browse/UILD-693
 [UILD-685]:https://folio-org.atlassian.net/browse/UILD-685
 [UILD-694]:https://folio-org.atlassian.net/browse/UILD-694
+[UILD-696]:https://folio-org.atlassian.net/browse/UILD-696
 
 ## 1.0.5 (2025-04-30)
 * Fixed incorrect behavior when navigating between duplicated resources. Fixes [UILD-553].